+++ src/app/components/Avatar/index.tsx
@@ -1,0 +1,3 @@
+import Avatar from '@material-ui/core/Avatar';
+
+export default Avatar;
