+++ src/app/components/Button/index.tsx
@@ -1,0 +1,3 @@
+import { Button } from '@material-ui/core';
+
+export default Button;
