+++ src/app/components/Card/Card.tsx
@@ -1,0 +1,3 @@
+import { Card } from '@material-ui/core';
+
+export { Card };
