+++ src/app/components/Card/CardActionArea.tsx
@@ -1,0 +1,3 @@
+import { CardActionArea } from '@material-ui/core';
+
+export { CardActionArea };
