+++ src/app/components/Card/CardActions.tsx
@@ -1,0 +1,3 @@
+import { CardActions } from '@material-ui/core';
+
+export { CardActions };
