+++ src/app/components/Card/CardContent.tsx
@@ -1,0 +1,3 @@
+import { CardContent } from '@material-ui/core';
+
+export { CardContent };
