+++ src/app/components/Card/CardMedia.tsx
@@ -1,0 +1,3 @@
+import { CardMedia } from '@material-ui/core';
+
+export { CardMedia };
