+++ src/app/components/Card/index.tsx
@@ -1,0 +1,7 @@
+import { Card } from './Card';
+import { CardActionArea } from './CardActionArea';
+import { CardContent } from './CardContent';
+import { CardMedia } from './CardMedia';
+import { CardActions } from './CardActions';
+
+export { Card, CardActionArea, CardContent, CardMedia, CardActions };
