+++ src/app/components/Checkbox/index.tsx
@@ -1,0 +1,3 @@
+import Checkbox from '@material-ui/core/Checkbox';
+
+export default Checkbox;
