+++ src/app/components/FormControlLabel/index.tsx
@@ -1,0 +1,3 @@
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+
+export default FormControlLabel;
