+++ src/app/components/Radio/Radio.tsx
@@ -1,0 +1,3 @@
+import { Radio } from '@material-ui/core';
+
+export default Radio;
