+++ src/app/components/Radio/RadioGroup.tsx
@@ -1,0 +1,3 @@
+import { RadioGroup } from '@material-ui/core';
+
+export default RadioGroup;
