+++ src/app/components/Typography/index.tsx
@@ -1,0 +1,3 @@
+import { Typography } from '@material-ui/core';
+
+export default Typography;
