--- src/app/components/index.ts
@@ -1,3 +1,0 @@
-import NavBar from './NavBar';
-
-export { NavBar };
