+++ src/app/pages/HomePage/Features/GithubRepoForm/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import styled from 'styled-components/macro';
 import { useSelector, useDispatch } from 'react-redux';
-import { FormLabel } from 'app/components/FormLabel';
+import { FormLabel } from 'app/components/backup/FormLabel';
 import { Input } from './components/Input';
 import { RepoItem } from './RepoItem';
 import { TextButton } from './components/TextButton';
@@ -11,7 +11,7 @@ import {
   selectLoading,
   selectError,
 } from './slice/selectors';
-import { LoadingIndicator } from 'app/components/LoadingIndicator';
+import { LoadingIndicator } from 'app/components/backup/LoadingIndicator';
 import { RepoErrorType } from './slice/types';
 import { useGithubRepoFormSlice } from './slice';
