+++ src/app/pages/HomePage/Features/ThemeSwitch/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { FormLabel } from 'app/components/FormLabel';
-import { Radio } from 'app/components/Radio';
+import { FormLabel } from 'app/components/backup/FormLabel';
+import { Radio } from 'app/components/backup/Radio';
 import styled from 'styled-components/macro';
 import { themeActions } from 'styles/theme/slice';
 import { useDispatch, useSelector } from 'react-redux';
