+++ src/app/pages/Login/index.tsx
@@ -1,16 +1,18 @@
 import React, { useState } from 'react';
-import Button from '@material-ui/core/Button';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import TextField from '@material-ui/core/TextField';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import Checkbox from '@material-ui/core/Checkbox';
 import Link from '@material-ui/core/Link';
 import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 import { Helmet } from 'react-helmet-async';
-import { NavBar } from 'app/components';
+import {
+  Button,
+  Checkbox,
+  FormControlLabel,
+  NavBar,
+  TextField,
+  Typography,
+} from 'app/components';
 import { connect } from 'react-redux';
 import { userActions } from 'actions';
 import { useHistory } from 'react-router-dom';
